+++ src/js/menu-info.js
@@ -1,4 +1,5 @@
 import $ from "jquery";
+import { printMenu } from "./menu-template";
 
 
 
@@ -29,7 +30,52 @@ function writeMenu(event){
 
 function writeReservations(event){
   $(".menu-info-container").html(`
-    This gonna be a form
+
+    <p class="control">
+      <input class="input is-small" type="text" placeholder="Small input">
+    </p>
+    <p class="control">
+      <input class="input" type="text" placeholder="Normal input">
+    </p>
+    <p class="control">
+      <input class="input is-medium" type="text" placeholder="Medium input">
+    </p>
+    <p class="control">
+      <input class="input is-large" type="text" placeholder="Large input">
+    </p>
+    <p class="control">
+      <span class="select is-small">
+        <select>
+          <option>Select dropdown</option>
+          <option>With options</option>
+        </select>
+      </span>
+    </p>
+    <p class="control">
+      <span class="select">
+        <select>
+          <option>Select dropdown</option>
+          <option>With options</option>
+        </select>
+      </span>
+    </p>
+    <p class="control">
+      <span class="select is-medium">
+        <select>
+          <option>Select dropdown</option>
+          <option>With options</option>
+        </select>
+      </span>
+    </p>
+    <p class="control">
+      <span class="select is-large">
+        <select>
+          <option>Select dropdown</option>
+          <option>With options</option>
+        </select>
+      </span>
+    </p>
+
   `);
 }
 
@@ -40,110 +86,5 @@ function getMenu () {
   });
 }
 
-function printMenu(data){
-  $(".menu-info-container").html(`
-    <div class="menu-title">Beer, Food, and Games</div>
-    <div class="divide-line1"></div>
-    <div class="menu-section-title">Beer</div>
-    <div class="divide-line2"></div>
-  `);
-  for(var count = data.Beer.length - 1; count >= 0; count--){
-    $(".divide-line2").after(`
-    <div class="menu-item-menu-price-container">
-      <div class="menu-item">${data.Beer[count].item}</div>............................
-      <div class="menu-item-price">${data.Beer[count].price}</div>
-    </div>
-    <div class="menu-item-style">${data.Beer[count].style}</div>
-    <div class="menu-description-allergies-favorite-container">
-      <div class="menu-item-description float-left">${data.Beer[count].description}</div>...
-      <div class="favorite${count} float-right"></div>
-      <div class="allergies${count} float-right"></div>
-    </div>
-    `);
-
-    if(data.Beer[count].allergies === 1){
-      $(`.allergies${count}`).html(`
-      <i class="float-right fa fa-exclamation-triangle" aria-hidden="true"></i>`);
-    }else{
-      $(`.allergies${count}`).html("");
-    }
-
-    if(data.Beer[count].favorite === 1){
-      $(`.favorite${count}`).html(`
-      <i class="float-right fa fa-star-o" aria-hidden="true"></i>`);
-    }else{
-      $(`.favorite${count}`).html("");
-    }
-  }
-
-  $(".menu-info-container").append(`
-    <div class="menu-section-title">Entrees</div>
-    <div class="divide-line3"></div>
-    `);
-  for(var i = data.entrees.length - 1; i>= 0; i--){
-    $(".divide-line3").after(`
-    <div class="menu-item-menu-price-container">
-      <div class="menu-item">${data.entrees[i].item}</div>............................
-      <div class="menu-item-price">${data.entrees[i].price}</div>
-    </div>
-
-    <div class="menu-description-allergies-favorite-container">
-      <div class="menu-item-description float-left">${data.entrees[i].description}</div>...
-      <div class="favorite1${i} float-right"></div>
-      <div class="allergies1${i} float-right"></div>
-    </div>
-  `);
-
-  if(data.entrees[i].allergies === 1){
-    $(`.allergies1${i}`).html(`
-    <i class="float-right fa fa-exclamation-triangle" aria-hidden="true"></i>`);
-  }else{
-    $(`.allergies1${i}`).html("");
-  }
-
-  if(data.entrees[i].favorite === 1){
-    $(`.favorite1${i}`).html(`
-    <i class="float-right fa fa-star-o" aria-hidden="true"></i>`);
-  }else{
-    $(`.favorite1${i}`).html("");
-  }
-}
-
-$(".menu-info-container").append(`
-  <div class="menu-section-title">Games</div>
-  <div class="divide-line4"></div>
-  `);
-for(var x = data.games.length - 1; x >= 0; x--){
-  $(".divide-line4").after(`
-  <div class="menu-item-menu-price-container">
-    <div class="menu-item">${data.games[x].item}</div>............................
-    <div class="menu-item-price">${data.games[x].price}</div>
-  </div>
-
-  <div class="menu-description-allergies-favorite-container">
-    <div class="menu-item-description float-left">${data.games[x].description}</div>...
-    <div class="favorite2${x} float-right"></div>
-    <div class="multi-player${x} float-right"></div>
-  </div>
-`);
-
-if(data.games[x].multiplayer === 1){
-  $(`.multi-player${x}`).html(`
-  <i class="fa fa-users" aria-hidden="true"></i>`);
-}else{
-  $(`.multi-player${x}`).html("");
-}
-
-if(data.games[x].favorite === 1){
-  $(`.favorite2${x}`).html(`
-  <i class="float-right fa fa-star-o" aria-hidden="true"></i>`);
-}else{
-  $(`.favorite2${x}`).html("");
-}
-}
-
-
-  console.log(data);
-}
 
 export { writeMenuInfo };
