+++ src/sass/josh.scss
@@ -1,3 +1,5 @@
+ @import "bulma";
+
 .our-story {
   display: inline-block;
   border: 1px solid gray;
